Changing app theme color
The background color of the app has been changed from a darker shade to a lighter, bluish tint (from `Color.fromARGB(18, 20, 25, 255)` to `Color.fromARGB(38, 10, 10, 255)`). This change is applied to both the home screen (`senpai_home.dart`) and the details screen (`senpai_details_screen.dart`).

Additionally, the text and icon colors on the details screen have been updated to a light blue (`Color(0xffdbe6ff)`) to maintain contrast and visibility with the new background. The app bar on the details screen now also includes a title "Details" with the same light blue text color.